move get ophysexperiment metadata to ophys_experiment
